requirements.txt
accelerate==1.7.0
datasets==3.6.0
faiss_cpu==1.11.0
Flask==3.1.1
matplotlib==3.10.3
mltu==1.2.5
numpy==2.2.6
opencv_python==4.11.0.86
path.py==12.5.0
pdf2image==1.17.0
peft==0.15.2
scikit_learn==1.6.1
sentence_transformers==4.1.0
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
ultralytics==8.3.146
Werkzeug==3.1.3
